example/requirements.txt
django-affiliate
Pillow
